+++ src/api/projects/projects.rs
@@ -147,6 +147,11 @@ pub struct Projects<'a> {
     /// Filter projects by those without activity before this date.
     #[builder(default)]
     last_activity_before: Option<DateTime<Utc>>,
+    /// Filter projects by which storage backend the repository is on.
+    ///
+    /// Available to administrators only.
+    #[builder(setter(into), default)]
+    repository_storage: Option<Cow<'a, str>>,
 
     /// Order results by a given key.
     #[builder(default)]
@@ -234,6 +239,7 @@ impl<'a> Endpoint for Projects<'a> {
             .push_opt("id_before", self.id_before)
             .push_opt("last_activity_after", self.last_activity_after)
             .push_opt("last_activity_before", self.last_activity_before)
+            .push_opt("repository_storage", self.repository_storage.as_ref())
             .extend(
                 self.custom_attributes
                     .iter()
@@ -595,6 +601,22 @@ mod tests {
 
         let endpoint = Projects::builder()
             .last_activity_before(Utc.ymd(2020, 1, 1).and_hms_milli(0, 0, 0, 0))
+            .build()
+            .unwrap();
+        api::ignore(endpoint).query(&client).unwrap();
+    }
+
+    #[test]
+    fn endpoint_repository_storage() {
+        let endpoint = ExpectedUrl::builder()
+            .endpoint("projects")
+            .add_query_params(&[("repository_storage", "default")])
+            .build()
+            .unwrap();
+        let client = SingleTestClient::new_raw(endpoint, "");
+
+        let endpoint = Projects::builder()
+            .repository_storage("default")
             .build()
             .unwrap();
         api::ignore(endpoint).query(&client).unwrap();